Update song.py
I registered a handler function to return a tuple with the file path, project name, and set name for any set that is currently open. I thought this would be helpful as I have wanted this feature in the past. I really love AbletonOSC by the way, thanks so much for maintaining it the way you all do!